+++ cmd/serve_windows_test.go
@@ -2,10 +2,6 @@ package cmd
 
 import "testing"
 
-func TestRunServe(t *testing.T) {
-	t.Skip("This test isn't implemented for Windows yet")
-}
-
 func TestRunServe_interrupt(t *testing.T) {
 	t.Skip("This test isn't implemented for Windows yet")
 }